pyproject.toml
[project]
name = "insult-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Ashwin Natarajan",email = "ashwin.nat@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "discord-py (>=2.5.2,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
